Fix UBSan warnings. (#1061)
* Fix UBSan warnings.

* Fix macOS build failures.

* Force NaN to 0 in fast_sqrt.

* Add PR #1061 to changelog.

+++ src/audio/MacAudioDevice.h
@@ -48,7 +48,7 @@ public:
 
     virtual bool isRunning() override;
     
-    virtual int getLatencyInMicroseconds() override;
+    virtual int64_t getLatencyInMicroseconds() override;
     
     // Configures current thread for real-time priority. This should be
     // called from the thread that will be operating on received audio.
